test: split vitest backend and browser mode
