+++ content/blog/fix-the-not-wrapped-in-act-warning/__tests__/timers.test.browser.jsx
