+++ package.json
@@ -42,8 +42,12 @@
     "pretest:e2e:run": "npm run build",
     "test:e2e:run": "cross-env CI=1 PORT=8811 playwright test",
     "test:e2e:install": "npx playwright install chromium --with-deps",
-    "test": "vitest --run",
-    "test:watch": "vitest",
+    "test": "run-s test:backend test:browser",
+    "test:backend": "vitest --run --config vitest.config.ts",
+    "test:backend:watch": "vitest --config vitest.config.ts",
+    "test:browser": "vitest --run --config vitest.browser.config.ts",
+    "test:browser:watch": "vitest --config vitest.browser.config.ts",
+    "test:watch": "npm run test:backend:watch",
     "lint": "eslint .",
     "format": "prettier --write \"**/*.+(js|jsx|json|yml|yaml|css|less|scss|ts|tsx|md|gql|graphql|mdx|vue)\"",
     "typecheck": "react-router typegen && tsc -b .",
@@ -194,6 +198,7 @@
     "@types/source-map-support": "^0.5.10",
     "@types/uuid": "^10.0.0",
     "@types/ws": "^8.18.1",
+    "@vitest/browser-playwright": "^4.0.18",
     "concurrently": "^9.2.1",
     "cookie": "^1.1.1",
     "esbuild-jest": "^0.5.0",
