+++ tests/msw-server.ts
@@ -1,0 +1,5 @@
+import { setupServer } from 'msw/node'
+import { mswHandlers } from '#mocks/msw-handlers.ts'
+
+export const mswServer = setupServer(...mswHandlers)
+
