+++ vitest.browser.config.ts
@@ -1,0 +1,22 @@
+/// <reference types="vitest" />
+/// <reference types="vite/client" />
+
+import react from '@vitejs/plugin-react'
+import { playwright } from '@vitest/browser-playwright'
+import tsconfigPaths from 'vite-tsconfig-paths'
+import { envOnlyMacros } from 'vite-env-only'
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+	plugins: [envOnlyMacros(), react(), tsconfigPaths()],
+	test: {
+		include: ['**/*.test.browser.{js,jsx,ts,tsx}'],
+		setupFiles: ['./tests/setup-browser.ts'],
+		browser: {
+			enabled: true,
+			provider: playwright(),
+			instances: [{ browser: 'chromium' }],
+		},
+	},
+})
+
