+++ vitest.config.ts
@@ -3,12 +3,14 @@
 
 import react from '@vitejs/plugin-react'
 import { defineConfig } from 'vitest/config'
+import tsconfigPaths from 'vite-tsconfig-paths'
 
 export default defineConfig({
-	plugins: [react()],
+	plugins: [react(), tsconfigPaths()],
 	test: {
 		include: ['**/__tests__/**.{js,jsx,ts,tsx}'],
-		environment: 'jsdom',
-		setupFiles: ['./tests/setup-env.ts'],
+		exclude: ['**/*.test.browser.{js,jsx,ts,tsx}'],
+		environment: 'node',
+		setupFiles: ['./tests/setup-backend.ts'],
 	},
 })
